Fixed enemy AI, moved some files for fomatting sake
I fixed the enemy movement script, you can test the Enemy Movement with a button attached to the script that resets the "hasmoved" boolean.

Furthermore, I added some basic shit for the CombatState script, it essentially handles most things turn-related. It's not fleshed out at all and we need to develop on that front

Also moved files for the PlayerUnitScript into it's own folder, scripts folder was getting messy. As well as the EnemySpawnerScript.