Player State Machine and MouseController update
Minor update to mouseController, it can now detect an enemyUnitScript on a gameObject and return it's reference to the cursor. This so that it can build on the functionality of showing enemyUnit details in the UI later.

Aside from that, a lot of the playerStates have their logic updated and finalized. I also added a new playerState script, it's called "isMoving",  this is to prevent mouse clicking while the player is moving, so that the player can't shift direction mid-movement.

GameEventSystem also got some stuff updated so that it can receive and acknowledge when a player has been selected.