WIP fixing click registration on entities
There's a pretty big bug in regards to how handling attacks and handling unit selected are being handled by the game in it's current state.

So I'm working towards reworking it to make it more consistent, since it's pretty shit in it's current state. It's not a huge rework and shouldn't necessarily break any other systems we have in place, given that I'm adding stuff rather than outright deleting it.