DeployPhase initial
Refactored some code to set up for the deploy phase. Currently I made it so that the CombatStateManager starts in the deploy phase.

So keep that in mind if you want to test stuff in this branch. You need to swap things to their specific combat states so that scripts related to those states can function properly.