Implemented state machine
Moved EnemyAI into state machine. Punching Bag Movement Script is unchanged so it can still be used for testing on other units.

Truth be told, using a scriptableObject for movement would be good as we can make custom movement scripts at that point.

Regardless, changes from the previous commit to the PunchingBagMovement were reverted and modified to fit the current implementation of the enemyMovement.

Don't worry, it doesn't break the old enemy prefab either. So it still works, luckily. Still working to fix the EnemyMovement in the state machine implementation.