Almost done, just need to refine turn controller script
Read above.

But for a more detailed description. Currently the script works by having the enemy turn end after a set amount of time.

This is fine for testing purposes but obviously not want we want in terms of gameplay. It's just one of those things I'll have to fix later.